requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
pytest-xdist
python-dotenv
